% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enet.R
\name{enet}
\alias{enet}
\title{Perform elastic net regression with scRNA-seq and spatial reference data}
\usage{
enet(
  x,
  y,
  adaptive = TRUE,
  hybrid = TRUE,
  tau = 1,
  nfolds = 10,
  n_run = 10,
  alpha = seq(0.2, 0.8, by = 0.1),
  ...
)
}
\arguments{
\item{x}{a feature by bins matrix used as predictors.}

\item{y}{a feature by cells matrix while each column will be used as response variables separately in separate lasso regressions.
The row names of y MUST be identical to those in x. If family = "gaussian" (default), it might be better to transform counts via logarithm (e.g., log2)}

\item{adaptive}{whether using the adaptive process (Default TRUE).}

\item{hybrid}{whether to elastic net or lasso (Default TRUE).}

\item{tau}{cofficient in transforming the weights to penalty as abs(weights)^(-tau). 1 by default, other choices could be 0.5, 2.}

\item{nfolds}{the number of folds used in cross validation to estimate lambda (lambda and alpha in elastic net).}

\item{n_run}{times of regresss to repeat per cell.}

\item{alpha}{list of alpha value used in evaluation. Note high alpha value tends to lasso regression.}

\item{...}{additional parameters to \code{glmnet::cv.glmnet} or \code{glmnetUtils::glmnetUtils}.}
}
\value{
A probility marix with bins in rownames and samples in columns, suggesting the probility of a sample assigned to a bin.
}
\description{
This function takes two input matrix, x and y, to  performs elastic net regression (via \code{glmnet}) on each column
of y (cells) with the x used as the predictors.
}
\details{
For each sample in the y, \code{enet} tries to find the weights of bins in x using non-negative elastic net. If \code{adaptive} is TRUE,
    ridge regression will be first performed to estimate the \code{penalty.factor}. If \code{hybrid} is TRUE,
    elastic net regression will be used with alpha and lambda are both estimated via cross validation.
    The estmatied coefficients are normalized by the sum, and averaged value from repeated runs are returned.
}
