% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_bin}
\alias{example_bin}
\title{Fly embyro spatial data}
\format{
A matrix
}
\usage{
data(example_bin)
}
\description{
A gene by bin matrix, values are quantile normalized.
}
\keyword{datasets}
