% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_expr}
\alias{example_expr}
\title{Fly embyro single-cell RNAseq data}
\format{
A sparse matrix
}
\usage{
data(example_expr)
}
\description{
A gene by cell matrix, values are normalized on read depth. See Nikos Karaiskos et al., (2017) for details.
}
\keyword{datasets}
