% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.R
\name{enet}
\alias{enet}
\title{Perform elastic net regression on scRNA-seq and spatial reference data}
\usage{
enet(
  x,
  y,
  adaptive = TRUE,
  hybrid = TRUE,
  tau = 1,
  nfolds = 10,
  n_run = 10,
  ...
)
}
\arguments{
\item{x}{a feature by sample matrix used as predictors.}

\item{y}{a feature by sample response matrix while each column will be used as response variables in separate lasso regressions.
The row names of y MUST be identical to those in x. If family = "gaussian", it might be better to transform counts via logarithm (e.g., log2)}

\item{adaptive}{whether using adaptive lasso (Default TRUE)}

\item{hybrid}{whether to elastic net (hybrid of lasso and ridge regression) or only lasso.}

\item{tau}{weights in transforming the penalty. 1 by default, other choices are 0.5, 2.}

\item{nfolds}{the number of folds used in cross validation to estimate lambda (and alpha if sepecified).}

\item{n_run}{number of regression to run per cell.}

\item{...}{additional parameters to \code{glmnet::cv.glmnet} or \code{glmnetUtils::glmnetUtils}.}
}
\value{
A probility marix with bins in rownames and samples in columns, suggesting the probility of a sample assigned to a bin.
}
\description{
This function takes two input matrix, x and y,  performs elastic net regression (via \code{glmnet}) on each column
of y with the x used the predictors.
}
\details{
For each sample in the y, this function tries to find the weights to the bins in x, with non-negative lasso.  If \code{adaptive} is TRUE,
ridge regression will be first perfomred to estimate the \code{penalty.factor}. Optionaly, the alpha value is estiamted via \code{glmnetUtils::cva.glmnet}.
lambda value will always be choosen from \code{nfolds} cross validation.
The estmatied coefficients are normalied to the sum, and the normalized value from \code{n_run} are averaged.
}
